rpi/requirements.txt
celery
flask
picamera2
python-decouple
requests
